add updated library structure + draft parallel solns
