+++ src/vegtamr/lic/__init__.py
@@ -1,0 +1,1 @@
+from ._api import compute_lic, compute_lic_with_postprocessing
